# Faster streaming of Labels to JSON