@bboreham 2023-07-25
Prombench shows a reduction in memory allocations, a very slight CPU reduction; not much else is different.

<img width="404" alt="image" src="https://github.com/prometheus/prometheus/assets/8125524/3f44c718-e25e-47da-9b6f-0493a1e1d541">
<img width="409" alt="image" src="https://github.com/prometheus/prometheus/assets/8125524/c37ceee3-e310-411b-ad3d-1f28c8e5f066">