@bboreham 2023-08-01
Note @alanprot approved this PR just above the prombench comment.
https://github.com/prometheus/prometheus/pull/12598#pullrequestreview-1544244732

> LGTM, I assume this is very similar to @alanprot just with the references in other marshallers (:

Thanks.  My PR also has a different benchmark.