@alanprot 2023-08-01
Yeah.

Thanks for looking into this @bboreham 

I closed https://github.com/prometheus/prometheus/pull/12598 in favor of this one